Moved workerType config funcs to new file and added validator
