+++ internal/cmdUtils/cmdUtils.go
@@ -449,40 +449,3 @@ func GetServiceConfigOverrideKeyOrGlobalKey(checkServiceConfigPath, key string) 
 	}
 	return
 }
-
-// GetWorkerConfigValue retrieves the value of a worker-specific key from the configuration
-func GetWorkerConfigValue(workerType, key string) any {
-	workerConfigPath := "workerType." + workerType + "." + key
-	return viper.Get(workerConfigPath)
-}
-
-// GetWorkerConfigString retrieves the configuration value for the given worker type and key,
-// and returns it as a string. If the value is not a string, an empty string is returned.
-func GetWorkerConfigString(workerType, key string) string {
-	val := GetWorkerConfigValue(workerType, key)
-	if v, ok := val.(string); ok {
-		return v
-	}
-	return ""
-}
-
-// GetWorkerConfigInt retrieves the configuration value for the given worker type and key,
-// and returns it as a string. If the value is not a string, an empty string is returned.
-func GetWorkerConfigInt(workerType, key string) int {
-	val := GetWorkerConfigValue(workerType, key)
-	if v, ok := val.(int); ok {
-		return v
-	}
-	return 0
-}
-
-// GetWorkerConfigStringSlice retrieves the configuration value for the given worker type and key,
-// and returns it as a slice of strings. If the value is not a []string, an empty slice is returned.
-func GetWorkerConfigStringSlice(workerType, key string) []string {
-	empty := make([]string, 0)
-	val := GetWorkerConfigValue(workerType, key)
-	if v, ok := val.([]string); ok {
-		return v
-	}
-	return empty
-}
